fix: TestFixtures 생성
