+++ src/test/java/H2QueryBuilder/H2QueryBuilderDDLTest.java
@@ -1,10 +1,10 @@
 package H2QueryBuilder;
 
 import common.ErrorCode;
-import jakarta.persistence.*;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import static H2QueryBuilder.fixtures.BuilderDDLFixtures.*;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
@@ -13,17 +13,10 @@ class H2QueryBuilderDDLTest {
     @Test
     void confirmIdAnnotationTest() {
         //given
-        @Entity
-        class Person {
-
-            @Id
-            private Long id;
-
-        }
         H2QueryBuilderDDL h2QueryBuilderDDL = new H2QueryBuilderDDL();
 
         //when
-        String createQuery = h2QueryBuilderDDL.create(Person.class);
+        String createQuery = h2QueryBuilderDDL.create(아이디만_존재하는_클래스());
 
         //then
         assertThat(createQuery).isEqualTo(
@@ -35,23 +28,10 @@ class H2QueryBuilderDDLTest {
     @Test
     void notExistColumnAnnotationTest() {
         //given
-        @Entity
-        class Person {
-
-            @Id
-            private Long id;
-
-            private String name;
-
-            private Integer age;
-
-            private String email;
-
-        }
         H2QueryBuilderDDL h2QueryBuilderDDL = new H2QueryBuilderDDL();
 
         //when
-        assertThat(h2QueryBuilderDDL.create(Person.class)).isEqualTo(
+        assertThat(h2QueryBuilderDDL.create(컬럼어노테이션_존재하지않는_클래스())).isEqualTo(
                 "CREATE TABLE Person (id BIGINT NOT NULL PRIMARY KEY, name VARCHAR, age INT, email VARCHAR);"
         );
     }
@@ -60,28 +40,10 @@ class H2QueryBuilderDDLTest {
     @Test
     void existGeneratedValueAnnotationOverTwoTest() {
         //given
-        @Table(name = "users")
-        @Entity
-        class Person {
-
-            @Id
-            @GeneratedValue(strategy = GenerationType.IDENTITY)
-            private Long id;
-
-            @Column(name = "nick_name")
-            private String name;
-
-            @Column(name = "old")
-            private Integer age;
-
-            @Column(nullable = false)
-            private String email;
-
-        }
         H2QueryBuilderDDL h2QueryBuilderDDL = new H2QueryBuilderDDL();
 
         //when, then
-        assertThat(h2QueryBuilderDDL.create(Person.class)).isEqualTo(
+        assertThat(h2QueryBuilderDDL.create(모든어노테이션_존재하는_클래스())).isEqualTo(
                 "CREATE TABLE users (id BIGINT NOT NULL AUTO_INCREMENT PRIMARY KEY, nick_name VARCHAR, old INT, email VARCHAR NOT NULL);"
         );
     }
@@ -90,28 +52,10 @@ class H2QueryBuilderDDLTest {
     @Test
     void createDropQueryTest() {
         //given
-        @Table(name = "users")
-        @Entity
-        class Person {
-
-            @Id
-            @GeneratedValue(strategy = GenerationType.IDENTITY)
-            private Long id;
-
-            @Column(name = "nick_name")
-            private String name;
-
-            @Column(name = "old")
-            private Integer age;
-
-            @Column(nullable = false)
-            private String email;
-
-        }
         H2QueryBuilderDDL h2QueryBuilderDDL = new H2QueryBuilderDDL();
 
         //when, then
-        assertThat(h2QueryBuilderDDL.drop(Person.class)).isEqualTo(
+        assertThat(h2QueryBuilderDDL.drop(모든어노테이션_존재하는_클래스())).isEqualTo(
                 "DROP TABLE users;"
         );
     }
@@ -120,29 +64,10 @@ class H2QueryBuilderDDLTest {
     @Test
     void notSupplySqlTypeTest() {
         //given
-        @Table(name = "users")
-        @Entity
-        class Person {
-            @Id
-            @GeneratedValue(strategy = GenerationType.IDENTITY)
-            private Long id;
-
-            @Column(name = "nick_name")
-            private String name;
-
-            @Column(name = "old")
-            private Integer age;
-
-            @Column(nullable = false)
-            private String email;
-
-            private Double cash;
-        }
-
         H2QueryBuilderDDL h2QueryBuilderDDL = new H2QueryBuilderDDL();
 
         //when, then
-        assertThatThrownBy(() -> h2QueryBuilderDDL.create(Person.class))
+        assertThatThrownBy(() -> h2QueryBuilderDDL.create(타입을_지원하지않는_클래스()))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessage(ErrorCode.NOT_ALLOWED_DATATYPE.getErrorMsg());
     }
